server/requirements.txt
asgiref==3.8.1
Django==5.1.4
sqlparse==0.5.3
psycopg2-binary==2.9.10
djangorestframework==3.15.2
boto3==1.36.2